Fixes related to Codemeta export. Test unification.
